fix(bootstrap): split store migrations to satisfy SLOC gate
Extract auth/config/discord migration helpers into
bootstrap_store_migrations.py so bootstrap_stores.py stays under
the 300 SLOC cap blocking PR #2024 CI.
